draw Z Position of the ball on canvas #72

+++ app/src/main/java/cz/fmo/Lib.java
@@ -51,7 +51,7 @@ public final class Lib {
         public int predecessorId; // unique identifier of the immediate predecessor
         public int centerX;       // midpoint of the object in the input image, X coordinate
         public int centerY;       // midpoint of the object in the input image, Y coordinate
-        public int centerZ;       // midpoint of the object in the input image, Z coordinate
+        public double centerZ;       // midpoint of the object in the input image, Z coordinate
         public boolean isBounce = false;
         public float directionX;  // unit orientation of the object, X coordinate
         public float directionY;  // unit orientation of the object, Y coordinate
